+++ jdaviz/configs/cubeviz/plugins/moment_maps/moment_maps.vue
@@ -23,16 +23,6 @@
     />
 
     <v-row>
-      <v-select
-        :items="viewers"
-        v-model="selected_viewer"
-        label='Plot in Viewer'
-        hint='Moment map will replace plot in the specified viewer.  Will also be available in the data dropdown in all image viewers.'
-        persistent-hint
-      ></v-select>
-    </v-row>
-
-    <v-row>
       <v-text-field
         ref="n_moment"
         type="number"
@@ -44,11 +34,19 @@
       ></v-text-field>
     </v-row>
 
-    <v-row justify="end">
-      <j-tooltip tipid='plugin-moment-maps-calculate'>
-        <v-btn color="primary" text @click="calculate_moment">Calculate</v-btn>
-      </j-tooltip>
-    </v-row>
+    <plugin-add-results
+      :label.sync="results_label"
+      :label_default="results_label_default"
+      :label_auto.sync="results_label_auto"
+      :label_invalid_msg="results_label_invalid_msg"
+      :label_overwrite="results_label_overwrite"
+      label_hint="Label for the collapsed cube"
+      :add_to_viewer_items="add_to_viewer_items"
+      :add_to_viewer_selected.sync="add_to_viewer_selected"
+      action_label="Calculate"
+      action_tooltip="Calculate moment map"
+      @click:action="calculate_moment"
+    ></plugin-add-results>
 
     <div v-if="moment_available">
       <j-plugin-section-header>Results</j-plugin-section-header>
@@ -56,7 +54,9 @@
           <v-text-field
            v-model="filename"
            label="Filename"
-           :rules="[() => !!filename || 'This field is required']">
+           hint="Export the latest calculated moment map"
+           :rules="[() => !!filename || 'This field is required']"
+           persistent-hint>
           </v-text-field>
       </v-row>
       <v-row justify="end">