+++ jdaviz/configs/default/plugins/gaussian_smooth/gaussian_smooth.vue
@@ -33,39 +33,23 @@
           type="number"
           hint="The stddev of the kernel, in pixels."
           persistent-hint
-          :rules="[() => !!stddev || 'This field is required', () => stddev > 0 || 'Kernel must be greater than zero']"
+          :rules="[() => !!stddev || 'This field is required',
+                   () => stddev > 0 || 'Kernel must be greater than zero']"
         ></v-text-field>
       </v-row>
 
-      <v-row v-if="dataset_selected && stddev > 0">
-        <v-select v-if="selected_mode == 'Spatial'"
-          :items="viewers"
-          v-model="selected_viewer"
-          label='Plot in Viewer'
-          hint='Smoothed data will replace plot in the specified viewer.  Will also be available in the data dropdown in all image viewers.'
-          persistent-hint></v-select>
-        <v-switch v-if="selected_mode == 'Spectral'"
-          label="Plot in Viewer"
-          hint="Smoothed data will immediately plot in the spectral viewer.  Will also be available in the data menu of each viewer."
-          v-model="add_replace_results"
-          persistent-hint>
-        </v-switch>
-      </v-row>
-
-      <v-row justify="end">
-        <j-tooltip v-if="selected_mode=='Spectral'" tipid='plugin-gaussian-apply'>
-          <v-btn :disabled="stddev <= 0 || dataset_selected == ''"
-            color="accent" text 
-            @click="spectral_smooth"
-          >Apply</v-btn>
-        </j-tooltip>
-        <j-tooltip v-else tipid='plugin-gaussian-apply'>
-          <v-btn 
-            :disabled="stddev <= 0 || dataset_selected == ''"
-            color="accent" text
-            @click="spatial_convolution"
-          >Apply</v-btn>
-        </j-tooltip>
-      </v-row>
+      <plugin-add-results
+        :label.sync="results_label"
+        :label_default="results_label_default"
+        :label_auto.sync="results_label_auto"
+        :label_invalid_msg="results_label_invalid_msg"
+        :label_overwrite="results_label_overwrite"
+        label_hint="Label for the smoothed data"
+        :add_to_viewer_items="add_to_viewer_items"
+        :add_to_viewer_selected.sync="add_to_viewer_selected"
+        action_label="Smooth"
+        action_tooltip="Smooth data"
+        @click:action="apply"
+      ></plugin-add-results>
     </j-tray-plugin>
 </template>