+++ jdaviz/configs/default/plugins/gaussian_smooth/tests/test_gaussian_smooth.py
@@ -17,12 +17,27 @@ def test_linking_after_spectral_smooth(spectrum1d_cube):
 
     gs = GaussianSmooth(app=app)
     gs.dataset_selected = 'test'
+    gs.selected_mode = 'Spectral'
     gs.stddev = 3.2
-    gs.add_replace_results = False
-    gs.vue_spectral_smooth()
+    gs.add_to_viewer_selected = 'None'
+    assert gs.results_label == 'spectral-smooth stddev-3.2'
+    gs.vue_apply()
+    # when not showing the results, the label will remain the same,
+    # so there should be an overwrite warning
+    assert gs.results_label_overwrite is True
+    gs.add_to_viewer_selected = 'spectrum-viewer'
+    gs.vue_apply()
+    # since we now plotted the results, the dataset should be fixed,
+    # but the dataset dropdown contains multiple choices, so the dataset
+    # itself is prepended to the default label, and there is no longer
+    # an overwrite warning.
+    assert len(gs.dataset_items) == 2
+    assert gs.dataset_selected == 'test'
+    assert gs.results_label == 'test spectral-smooth stddev-3.2'
+    assert gs.results_label_overwrite is False
 
     assert len(dc) == 2
-    assert dc[1].label == 'Smoothed test stddev 3.2'
+    assert dc[1].label == 'spectral-smooth stddev-3.2'
     assert len(dc.external_links) == 3
 
     assert dc.external_links[2].cids1[0] is dc[0].world_component_ids[0]
@@ -39,10 +54,12 @@ def test_spatial_convolution(spectrum1d_cube):
 
     gs = GaussianSmooth(app=app)
     gs.dataset_selected = 'test'
+    gs.selected_mode = 'Spatial'
     gs.stddev = 3
-    gs.vue_spatial_convolution()
+    assert gs.results_label == 'spatial-smooth stddev-3.0'
+    gs.vue_apply()
 
     assert len(dc) == 2
-    assert dc[1].label == "Smoothed test spatial stddev 3.0"
-    assert (dc["Smoothed test spatial stddev 3.0"].get_object(cls=Spectrum1D, statistic=None).shape
+    assert dc[1].label == "spatial-smooth stddev-3.0"
+    assert (dc["spatial-smooth stddev-3.0"].get_object(cls=Spectrum1D, statistic=None).shape
             == (4, 2, 2))