+++ CHANGES.rst
@@ -64,9 +64,12 @@ Other Changes and Additions
 - Data dropdowns in plugins are now filtered to only applicable entries. [#1221]
 - Cube data now has spectral axis last in the backend, to match specutils Spectrum1D
   axis order and work with updated glue-astronomy translators. [#1174]
-
+- Plugins that create data entries allow overriding the default labels. [#1239]
+- Automatic defaults for model component IDs and equation editor in model fitting. [#1239]
 - Help button in toolbar to open docs in a new tab. [#1240]
 - Snackbar queue handles loading interrupt more cleanly. [#1249]
+
+- Reported quantities are rounded/truncated to avoid showing unnecessary precision. [#1244]
 
 2.4 (2022-03-29)
 ================